+++ FYProject/FYProject/main.cpp
@@ -15,7 +15,7 @@ void printShortestPath(vector<int> sp) {
         cout << "There is no shortest path" << endl;
     }
     else {
-        for (int j = 0; j < sp.size(); j++)
+        for (size_t j = 0; j < sp.size(); j++)
         {
             cout << sp[j] << " --> ";
         }
